# Optimize dictFind by leveraging key length functions to avoid redundant computations.