--- a/src/dict.c
+++ b/src/dict.c
@@ -73,12 +73,26 @@ static int dictDefaultCompare(dict *d, const void *key1, const void *key2);
 /* -------------------------- misc inline functions -------------------------------- */
 
 typedef int (*keyCmpFunc)(dict *d, const void *key1, const void *key2);
+typedef size_t (*keyLenFunc)(dict *d, const void *key1);
+typedef int (*keyCmpFuncWithLen)(dict *d, const void *key1, const void *key2, const size_t key1Len,const size_t key2Len);
 static inline keyCmpFunc dictGetKeyCmpFunc(dict *d) {
     if (d->useStoredKeyApi && d->type->storedKeyCompare)
         return d->type->storedKeyCompare;
     if (d->type->keyCompare)
         return d->type->keyCompare;
     return dictDefaultCompare;
+}
+
+static inline keyCmpFuncWithLen dictGetKeyCmpFuncWithLen(dict *d) {
+    if (d->type->keyCompareWithLen)
+        return d->type->keyCompareWithLen;
+    return NULL;
+}
+
+static inline keyLenFunc dictGetKeyLenFunc(dict *d) {
+    if (d->type->keyLen)
+        return d->type->keyLen;
+    return NULL;
 }
 
 static inline uint64_t dictHashKey(dict *d, const void *key, int isStoredKey) {
@@ -778,6 +792,11 @@ dictEntry *dictFindByHash(dict *d, const void *key, const uint64_t hash) {
     /* Rehash the hash table if needed */
     _dictRehashStepIfNeeded(d,idx);
 
+    /* Check if we can use the compare function with length to avoid recomputing length of key always */
+    keyCmpFuncWithLen cmpFuncWithLen = dictGetKeyCmpFuncWithLen(d);
+    keyLenFunc keyLenFunc = dictGetKeyLenFunc(d);
+    const int has_len_fn = (keyLenFunc!=NULL && cmpFuncWithLen != NULL);
+    const size_t keyLen = has_len_fn ? keyLenFunc(d,key) : 0;
     for (table = 0; table <= 1; table++) {
         if (table == 0 && (long)idx < d->rehashidx) continue;
         idx = hash & DICTHT_SIZE_MASK(d->ht_size_exp[table]);
@@ -791,8 +810,8 @@ dictEntry *dictFindByHash(dict *d, const void *key, const uint64_t hash) {
 
             /* Prefetch the next entry to improve cache efficiency */
             redis_prefetch_read(dictGetNext(he));
-
-            if (key == he_key || cmpFunc(d, key, he_key))
+            if (key == he_key ||
+                has_len_fn ? cmpFuncWithLen(d, key, he_key, keyLen, (keyLenFunc(d,he_key))): cmpFunc(d, key, he_key))
                 return he;
             he = dictGetNext(he);
         }