--- a/src/dict.c
+++ b/src/dict.c
@@ -809,7 +809,7 @@ dictEntry *dictFindByHash(dict *d, const void *key, const uint64_t hash) {
             /* Prefetch the next entry to improve cache efficiency */
             redis_prefetch_read(dictGetNext(he));
             if (key == he_key ||
-                has_len_fn ? cmpFuncWithLen(d, key, he_key, keyLen, (keyLenFunc(d,he_key))): cmpFunc(d, key, he_key))
+                has_len_fn ? cmpFuncWithLen(d, key, keyLen, he_key, (keyLenFunc(d,he_key))): cmpFunc(d, key, he_key))
                 return he;
             he = dictGetNext(he);
         }