--- a/src/dict.c
+++ b/src/dict.c
@@ -807,7 +807,7 @@ dictEntry *dictFindByHash(dict *d, const void *key, const uint64_t hash) {
             /* Prefetch the next entry to improve cache efficiency */
             redis_prefetch_read(dictGetNext(he));
             if (key == he_key ||
-                has_len_fn ? cmpFuncWithLen(d, key, key_len, he_key, (keyLenFunc(d,he_key))): cmpFunc(d, key, he_key))
+                (has_len_fn ? cmpFuncWithLen(d, key, key_len, he_key, (keyLenFunc(d,he_key))): cmpFunc(d, key, he_key)))
                 return he;
             he = dictGetNext(he);
         }