--- a/src/dict.h
+++ b/src/dict.h
@@ -93,9 +93,9 @@ typedef struct dictType {
     /* Optional callback called when the dict is destroyed. */
     void (*onDictRelease)(dict *d);
 
-    /* Optional keylen to avoid duplication computation of key lengths*/
-    size_t (*keyLen)(dict *d, const void *key);
-    int (*keyCompareWithLen)(dict *d, const void *key1, const size_t l1, const void *key2, const size_t l2);
+    /* Optional keylen to avoid duplication computation of key lengths. */
+    keyLenFunc keyLen;
+    keyCmpFuncWithLen keyCompareWithLen;
 } dictType;
 
 #define DICTHT_SIZE(exp) ((exp) == -1 ? 0 : (unsigned long)1<<(exp))