--- a/src/server.c
+++ b/src/server.c
@@ -282,8 +282,8 @@ size_t dictSdsLen(dict *d, const void *key) {
     return sdslen((sds)key);
 }
 
-int dictSdsKeyCompareWithLen(dict *d, const void *key1,
-        const void *key2, const size_t l1, const size_t l2)
+int dictSdsKeyCompareWithLen(dict *d, const void *key1, const size_t l1,
+        const void *key2,  const size_t l2)
 {
     UNUSED(d);
     if (l1 != l2) return 0;