This PR enhances dictFind by introducing support for key length functions, allowing the use of keyCompareWithLen when available. This avoids redundant key length computations, improving efficiency, especially when the dictionary is rehashing or there are a significant number of hash collisions.

Additionally, it maintains backward compatibility and optimizes key lookups without altering existing behavior.


## Performance improvement on 100% GETs use-case

benchmark command used
```
taskset -c 1-11 memtier_benchmark --ratio 0:1 --key-maximum 1000000 --key-minimum 1 -c 1 -t 5 --pipeline 100 --key-pattern P:P --test-time 30 --hide-histogram -d 1024 -S /tmp/1.socket  -x 3
```


In unstable dictFindByHash takes 29% (and sdslen within it takes 8.9%) of CPU cycles for a high-pipeline 100% gets use-case. 
After this change  dictFindByHash takes 27.8% (and sdslen within it takes 7.7%)

Function | CPU Time: Total: Difference | CPU Time: Total: 1cd622bdcabf1bd80fd20a7c5722808108e324ff | CPU Time: Total: this PR bea71c97b3ec31930a5f05b135393b5c29afce50 | CPU Time: Self: Difference | CPU Time: Self: 1cd622bdcabf1bd80fd20a7c5722808108e324ff | CPU Time: Self: this PR bea71c97b3ec31930a5f05b135393b5c29afce50 | Module | Function (Full) | Source File
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
sdslen | 1.220s | 8.90% | 7.70% | 1.220s | 8.888s | 7.669s | redis-server | sdslen | sds.h
dictFindByHash | 1.199s | 29.00% | 27.80% | 0.189s | 9.958s | 9.769s | redis-server | dictFindByHash | dict.c

![image](https://github.com/user-attachments/assets/7f80c4ee-41ee-4c93-8b45-8982878d7898)


Checks:
- [x] daily: https://github.com/filipecosta90/redis/actions/runs/13395593954