@fcostaoliveira 2025-02-07
### CE Performance Automation : step 1 of 2 (build) DONE.

This comment was automatically generated given a benchmark was triggered.
Started building at 2025-02-07 18:10:26.247727 and took 64 seconds.
You can check each build/benchmark progress in grafana:
   - git hash: bea71c97b3ec31930a5f05b135393b5c29afce50
   - git branch: filipecosta90:dict.compare.optimize
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*

You can check a comparison in detail via the [grafana link](https://benchmarksredisio.grafana.net/d/edsxdsrbexhc0f/ce-benchmark-run-status?orgId=1&var-benchmark_work_stream=1738951890708-0)